Social Authentication using Passport-Google-oAuth

+++ controllers/api/v1/posts_api.js
@@ -25,8 +25,8 @@ module.exports.destroy = async function(req,res)
 {
     try {
         let post = await Post.findById(req.params.id);
-        // if(post.user == req.user.id)
-        // {
+        if(post.user == req.user.id)
+        {
             post.remove();
             await Comment.deleteMany({post:req.params.id});
             
@@ -34,16 +34,17 @@ module.exports.destroy = async function(req,res)
             return res.json(200,{
                 message:"Post and associated comments deleted successfully"
             });
-        // }else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }else{
+            return res.status(401).json({
+                message:"You cannot delete this post"
+                });
+            }
 
-    } catch (error) {
+        } catch (error) {
         
         console.log('**** Error:',error);
         return res.status(500).json({
             message:'Internal Server Error'
-        });
-    }
+            });
+        }
 }
